moved the Stack.py code into realStack.py file
This moved the Stack.py code into realStack.py with the intention of renaming realStack.py to Stack.py. Additionally, any last issues involving the Stack class should be resolved here.
